# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for IAM and Mortgage (French Method) bounded contexts",
        "title": "Finanzas API - MiVivienda Mortgage Calculator",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/v1/iam/login": {
            "post": {
                "description": "Authenticate user and return token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IAM"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.LoginResource"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resources.LoginResponseResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/iam/register": {
            "post": {
                "description": "Register a new user with email, password, and full name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "IAM"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.RegisterUserResource"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/resources.UserResource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mortgage/calculate": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Calculates a mortgage loan using the French amortization method (constant installments)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortgage"
                ],
                "summary": "Calculate mortgage with French method",
                "parameters": [
                    {
                        "description": "Mortgage calculation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/resources.CalculateMortgageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resources.MortgageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mortgage/history": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get mortgage calculation history for a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortgage"
                ],
                "summary": "Get mortgage calculation history",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 50,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/resources.MortgageSummaryResource"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/mortgage/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a specific mortgage calculation by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mortgage"
                ],
                "summary": "Get mortgage by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Mortgage ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/resources.MortgageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "resources.CalculateMortgageRequest": {
            "type": "object",
            "required": [
                "currency",
                "down_payment",
                "grace_period_type",
                "interest_rate",
                "loan_amount",
                "property_price",
                "rate_type",
                "term_months",
                "user_id"
            ],
            "properties": {
                "bono_techo_propio": {
                    "type": "number",
                    "minimum": 0
                },
                "currency": {
                    "type": "string",
                    "enum": [
                        "PEN",
                        "USD"
                    ]
                },
                "down_payment": {
                    "type": "number",
                    "minimum": 0
                },
                "grace_period_months": {
                    "type": "integer",
                    "minimum": 0
                },
                "grace_period_type": {
                    "type": "string",
                    "enum": [
                        "NONE",
                        "TOTAL",
                        "PARTIAL"
                    ]
                },
                "interest_rate": {
                    "type": "number",
                    "minimum": 0
                },
                "loan_amount": {
                    "type": "number"
                },
                "npv_discount_rate": {
                    "type": "number",
                    "minimum": 0
                },
                "property_price": {
                    "type": "number"
                },
                "rate_type": {
                    "type": "string",
                    "enum": [
                        "NOMINAL",
                        "EFFECTIVE"
                    ]
                },
                "term_months": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "resources.LoginResource": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "mypassword123"
                }
            }
        },
        "resources.LoginResponseResource": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string",
                    "example": "user-1-token"
                },
                "user": {
                    "$ref": "#/definitions/resources.UserResource"
                }
            }
        },
        "resources.MortgageResponse": {
            "type": "object",
            "properties": {
                "bono_techo_propio": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "down_payment": {
                    "type": "number"
                },
                "fixed_installment": {
                    "type": "number"
                },
                "grace_period_months": {
                    "type": "integer"
                },
                "grace_period_type": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "interest_rate": {
                    "type": "number"
                },
                "irr": {
                    "type": "number"
                },
                "loan_amount": {
                    "type": "number"
                },
                "npv": {
                    "type": "number"
                },
                "payment_schedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/resources.PaymentScheduleItemResource"
                    }
                },
                "periodic_rate": {
                    "type": "number"
                },
                "principal_financed": {
                    "description": "Resultados calculados",
                    "type": "number"
                },
                "property_price": {
                    "type": "number"
                },
                "rate_type": {
                    "type": "string"
                },
                "tcea": {
                    "type": "number"
                },
                "term_months": {
                    "type": "integer"
                },
                "total_interest_paid": {
                    "type": "number"
                },
                "total_paid": {
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "resources.MortgageSummaryResource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "fixed_installment": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "loan_amount": {
                    "type": "number"
                },
                "property_price": {
                    "type": "number"
                },
                "tcea": {
                    "type": "number"
                },
                "term_months": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "resources.PaymentScheduleItemResource": {
            "type": "object",
            "properties": {
                "amortization": {
                    "type": "number"
                },
                "installment": {
                    "type": "number"
                },
                "interest": {
                    "type": "number"
                },
                "is_grace_period": {
                    "type": "boolean"
                },
                "period": {
                    "type": "integer"
                },
                "remaining_balance": {
                    "type": "number"
                }
            }
        },
        "resources.RegisterUserResource": {
            "type": "object",
            "required": [
                "email",
                "full_name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "full_name": {
                    "type": "string",
                    "minLength": 2,
                    "example": "John Doe"
                },
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "mypassword123"
                }
            }
        },
        "resources.UserResource": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "email": {
                    "type": "string",
                    "example": "user@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}